move "faster" tests away from "simple" tests
